pyproject.toml
[project]
name = "osdu-python-client"
version = "0.1.0"
description = "Python client for OSDU services"
readme = "README.md"
requires-python = ">=3.13"
license = "Apache-2.0"
authors = [
    {name = "OSDU team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

dependencies = [
    "attrs>=25.4.0",
    "httpx>=0.28.1",
    "openapi-python-client>=0.28.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "ruff>=0.14.14",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
managed = true